Layout Fixes
Still have to figure out some issues, not really digging the body size
at 62.5% to make REM units easier to work with. This causes the font
size to mess with the grids, while it could be fixed for layout with not
too much effort, it could be a giant pain once you start throwing media
queries on other random elements with a REM size set on the container.