Obscure Fixes
Sorting through some fixes through I noticed when stress testing things.
For example disabling all the layouts gives a .nolayout class, which I
tacked on to the content-sidebar layout, so there is never not a
layout... Like should never matter, but might as well fix stuff when you
notice it. :D